Realizamos ajustes en el Controlador de Ventas y Detalle de Ventas

+++ controllers/saleDetail.js
@@ -16,9 +16,23 @@ const getSalesDetails = async ( req, res = response ) => {
 
 }
 
+const getDetailBySaleId = async ( req, res = response ) => {
+
+    const saleId = req.params.id;
+
+    const saleDetails = await SaleDetail.find({ sale: saleId })
+                                            .populate( 'product', 'name salePrice' );
+
+    return res.status( 200 ).json({
+        ok: true,
+        saleDetails
+    });
+
+}
+
 const createSaleDetail = async ( req, res = response ) => {
 
-    const { sale, product } = req.body;
+    const { sale, product, quantity } = req.body;
 
     try {
         
@@ -42,9 +56,11 @@ const createSaleDetail = async ( req, res = response ) => {
         
         const savedSaleDetail = await newSaleDetail.save();
 
+        await Product.findByIdAndUpdate( product, { $set: { stock: productById.stock - quantity } } );
+
         return res.status( 201 ).json({
             ok: true,
-            savedSaleDetail
+            saleDetail: savedSaleDetail
         });
 
     } catch (error) {
@@ -65,15 +81,23 @@ const deleteSaleDetail = async ( req, res = response ) => {
 
     try {
         
-        const saleDetail = await SaleDetail.findById( saleDetailId ); //* buscar si venta existe
-        
+        const saleDetail = await SaleDetail.findById( saleDetailId ); //* buscar si detalle de venta existe
         if ( !saleDetail )
             return res.status( 400 ).json({
                 ok: false,
                 msg: 'El Id de detalle venta no existe'
             });
 
+        const productById = await Product.findById( saleDetail.product ); //* buscar si producto existe  
+        if ( !productById )
+            return res.status( 400 ).json({
+                ok: false,
+                msg: 'El Id de producto no existe'
+            }); 
+
         const deletedSaleDetail = await SaleDetail.findByIdAndDelete( saleDetailId );
+
+        await Product.findByIdAndUpdate( saleDetail.product, { $set: { stock: productById.stock + saleDetail.quantity } } );
 
         return res.status( 200 ).json({
             ok: true,
@@ -92,8 +116,43 @@ const deleteSaleDetail = async ( req, res = response ) => {
 
 }
 
+// const deleteSaleDetail = async ( req, res = response ) => {
+
+//     const saleId = req.params.id;
+
+//     try {
+        
+//         const saleById = await Sale.findById( saleId ); //* buscar si venta existe
+    
+//         if ( !saleById )
+//             return res.status( 400 ).json({
+//                 ok: false,
+//                 msg: 'El Id de venta no existe'
+//             });
+
+//         const deletedSaleDetail = await SaleDetail.deleteMany( { sale: saleId } );
+
+//         console.log( deletedSaleDetail );
+//         return res.status( 200 ).json({
+//             ok: true,
+//             deletedSaleDetail
+//         });
+
+//     } catch (error) {
+        
+//         console.log( error );
+//         return res.status( 500 ).json({
+//             ok: false,
+//             msg: 'Hable con el Administrador'
+//         });
+
+//     }
+
+// }
+
 module.exports = {
     getSalesDetails,
+    getDetailBySaleId,
     createSaleDetail,
     deleteSaleDetail
 }
